Create image2.cpp based on image_cross_cor.cpp
As before calculate the cross-correlation between two images.  However,
now there is C++ code (findCPPMax) to locate the offset that produces peak
correlation.  More importantly, added calcDisparityMap which introduces
the idea of iterating through smaller subsections of the image in a DFS
to produce a best offset for small groups of pixels instead of only one
offset for the entire image pair.  The distances represented by the
offsets then are plotted as a disparity map. This disparity map is what
is now returned as the output instead of the single correlation between
the two whole images.