Fix problem of finding incorrect matches in forested areas.
In findOffset find the max of conv - gaussian_blur(image, 91).
This filters out large "bright" areas and focuses the search
on the smaller (few pixel sized) peak that represents the real
match.  Large bright areas occur around edges (small offset)
of the correlation image because the forest is self similar.
But this manefests itself as a large blob centered at 0,0
and sometimes running the whole height of the image.  By
subtracting a blurred image, it normalizes it so that it
is sensitive to finding a peak that is significant relative
to the local area rather than just finding the one peak.